introduce ElectrodeCoreActivity
